requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
mysql-connector-python==9.4.0
pillow==12.0.0
sqlparse==0.5.3
